AI_ML/chatbot/requirements.txt
openai==0.28
python-dotenv
google-genai
streamlit>=1.28.0